+++ CHANGELOG.md
@@ -25,6 +25,7 @@ IMPROVEMENTS:
 * builder/digitalocean: Retry on any pending event errors.
 * builder/openstack: Can now specify a project. [GH-382]
 * builder/virtualbox: Can now attach hard drive over SATA. [GH-391]
+* provisioner/file: Can now upload directories. [GH-251]
 
 BUG FIXES: